# Fix silent fallback to Debug when building native libs