* Currently the CMAKE_BUILD_TYPE is not being passed in when
  creating the native library intermediates.
* Before the change to Ninja for CMake generation (8c2158f9fe1)
  the default uninitialized cmake build type was set to Release so the
  bug was not visible.
* After the Ninja change, the default uninitialized cmake build type
  is set to Debug so it becomes more apparent.
* This improves performance of native libs like System.IO.Compression
  by a considerable amount as they are currently being built in Debug,
  even when the Release configuration is specified.