@L2 2021-04-14
Example of the perf improvement, before and after this PR:

dotnet/performance: --filter: System.IO.Compression.Brotli.Decompress_WithoutState*

summary:
better: 6, geomean: 17.438
total diff: 6

No Slower results for the provided threshold = 2% and noise filter = 25ns.

| Faster                                                                                         | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.IO.Compression.Brotli.Decompress_WithoutState(level: Fastest, file: "TestDocument.pdf") |     33.32 |      10093487.50 |        302940.38 |         |
| System.IO.Compression.Brotli.Decompress_WithoutState(level: Fastest, file: "sum")              |     20.87 |       2377653.57 |        113907.56 |         |
| System.IO.Compression.Brotli.Decompress_WithoutState(level: Fastest, file: "alice29.txt")      |     19.69 |      10137654.69 |        514768.29 |         |
| System.IO.Compression.Brotli.Decompress_WithoutState(level: Optimal, file: "alice29.txt")      |     14.49 |       6833235.42 |        471663.54 |         |
| System.IO.Compression.Brotli.Decompress_WithoutState(level: Optimal, file: "sum")              |     13.36 |       1810263.19 |        135501.35 |         |
| System.IO.Compression.Brotli.Decompress_WithoutState(level: Optimal, file: "TestDocument.pdf") |     10.61 |       9129109.38 |        860791.32 |         |